+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.8.2] - 2026-07-09
+
+### Changed
+
+- Keep entity collections on the UI thread and queue console and trace updates in UI batches
+- Updated Avalonia and supporting packages to their latest compatible patch versions
+
 ## [1.8.1] - 2025-11-14
 
 ### Changed